+++ src/main/scala/com/github/plume/oss/Jimple2Cpg.scala
@@ -148,8 +148,12 @@ class Jimple2Cpg {
 
       basePasses(cpg, driver, unchangedTypes, unchangedNamespaces).foreach(_.createAndApply(driver))
       controlFlowPasses(cpg).foreach(_.createAndApply(driver))
-      new PlumeReachingDefPass(cpg).createAndApply(driver)
+      new PlumeReachingDefPass(cpg, unchangedTypes = unchangedTypes).createAndApply(driver)
       new PlumeHashPass(cpg).createAndApply(driver)
+      driver match {
+        case x: OverflowDbDriver => x.removeExpiredPathsFromCache(unchangedTypes)
+        case _                   =>
+      }
 
       driver.buildInterproceduralEdges()
       cpg